+++ plotly/tests/test_core/test_session/test_session.py
@@ -1,13 +1,12 @@
 from __future__ import absolute_import
 
-from unittest import TestCase
+from plotly.tests.utils import PlotlyTestCase
 
-from plotly.session import (update_session_plot_options, SHARING_OPTIONS,
-                            _session)
+from plotly.session import update_session_plot_options, SHARING_OPTIONS
 from plotly.exceptions import PlotlyError
 
 
-class TestSession(TestCase):
+class TestSession(PlotlyTestCase):
 
     def test_update_session_plot_options_invalid_sharing_argument(self):
 
@@ -22,8 +21,9 @@ class TestSession(TestCase):
         # _session['plot_options'] should contain sharing key after
         # update_session_plot_options is called by correct arguments
         # 'public, 'private' or 'secret'
-
+        from plotly.session import _session
         for key in SHARING_OPTIONS:
             kwargs = {'sharing': key}
             update_session_plot_options(**kwargs)
+
             self.assertEqual(_session['plot_options'], kwargs)