+++ plotly/tests/test_core/test_stream/test_stream.py
@@ -34,6 +34,7 @@ def test_initialize_stream_plot():
     stream = Stream(token=tk, maxpoints=50)
     url = py.plot([Scatter(x=[], y=[], mode='markers', stream=stream)],
                   auto_open=False,
+                  world_readable=True,
                   filename='stream-test')
     assert url == 'https://plot.ly/~PythonAPI/461'
     time.sleep(.5)
@@ -45,6 +46,7 @@ def test_stream_single_points():
     stream = Stream(token=tk, maxpoints=50)
     res = py.plot([Scatter(x=[], y=[], mode='markers', stream=stream)],
                   auto_open=False,
+                  world_readable=True,
                   filename='stream-test')
     time.sleep(.5)
     my_stream = py.Stream(tk)
@@ -60,6 +62,7 @@ def test_stream_multiple_points():
     stream = Stream(token=tk, maxpoints=50)
     url = py.plot([Scatter(x=[], y=[], mode='markers', stream=stream)],
                   auto_open=False,
+                  world_readable=True,
                   filename='stream-test')
     time.sleep(.5)
     my_stream = py.Stream(tk)
@@ -75,6 +78,7 @@ def test_stream_layout():
     stream = Stream(token=tk, maxpoints=50)
     url = py.plot([Scatter(x=[], y=[], mode='markers', stream=stream)],
                   auto_open=False,
+                  world_readable=True,
                   filename='stream-test')
     time.sleep(.5)
     title_0 = "some title i picked first"